// pwa/src/generated/models/attachmentResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wetty Chat API
 * Real-time chat application backend API supporting groups, messaging, threads, stickers, invites, and push notifications.
 * OpenAPI spec version: 0.1.0
 */
import type { SnowflakeID } from './snowflakeID';

export interface AttachmentResponse {
  fileName: string;
  height?: number;
  id: SnowflakeID;
  kind: string;
  size: number;
  url: string;
  width?: number;
}